Generated commit to update templated files based on rev a89f13e in stackabletech/operator-templating repo.
Original commit message:
Make copying of config-spec conditional (#62)

Some operators (looking at you, Regorule) do not have config-spec files, which caused the generate manifest step to fail after PR #61 which changed the syntax for copying these files.

This adds a check if any config-spec files are present and skips copying them if not. This requires that operators do not have .dummy files in place, as was the case for regorule (fixed in https://github.com/stackabletech/regorule-operator/pull/195 )