chore: Version CRD (#661)
* chore: migrate crd crate to operator-binary module

* chore: add stackable-versioned dependency

* chore: separate impls from crd data structures

This makes way for the versioned module we will soon introduce

* chore: version the crd

At this point, errors will appear in any crates using the crd.
It has only been done separately to illustrate the ease in versioning a CRD without all of the other necessary changes.

* chore: version any other structures defined outside of the main crd

This is helpful for later crd version sharing substructures that might not change.
For example: v1alpha2::OpaCluster might still use user_info_fetcher::v1alpha1::Config,
or perhaps it uses user_info_fetcher::v1beta1::Config.

Similarly, shared structures from stackable-operators can then be
versioned in the same way.

* chore: update references in the crates

The versioned module is imported rather than the individual structs and enums (when there is no conflict, eg: if also importing a versioned shared struct) so that that usages show the version explicitly.

There might be times where this isn't possible, for example, once structs and enums are versioned in stackable-operator, there could be multiple modules with the same name.

In this case, user-info-fetcher is also versioned with v1alpha1, so it is referred to as user_info_fetcher::v1alpha1 in crd/mod.rs so as to not conflict with the crds v1alpha1.

* chore: add attribution

Co-authored-by: Techassi <git@techassi.dev>

* docs: Update references to include the version

---------

Co-authored-by: Techassi <git@techassi.dev>